# Conflicts:
#	backend/models/holding.py

+++ backend/models/holding.py
@@ -5,13 +5,13 @@ from typing import Any, Self, Optional
 @dataclass
 class Holding:
     symbol: str
-    units: int
+    units: float
     property_metadata: Optional[dict[str, Any]] = None
 
     @classmethod
     def from_dict(cls, data: dict[str, Any]) -> Self:
         return cls(
             symbol=data["symbol"],
-            units=int(data["units"]),
+            units=float(data["units"]),
             property_metadata=data.get("property_metadata"),
         )
